+++ src/openpose/core/keyPointScaler.cpp
@@ -1,6 +1,6 @@
-#include "openpose/core/scaleKeyPoints.hpp"
-#include "openpose/utilities/errorAndLog.hpp"
-#include "openpose/core/keyPointScaler.hpp"
+#include <openpose/core/scaleKeyPoints.hpp>
+#include <openpose/utilities/errorAndLog.hpp>
+#include <openpose/core/keyPointScaler.hpp>
 
 namespace op
 {
@@ -9,7 +9,7 @@ namespace op
     {
     }
 
-    void KeyPointScaler::scale(Array<float>& arrayToScale, const double scaleInputToOutput, const double scaleNetToOutput, const cv::Size& producerSize) const
+    void KeyPointScaler::scale(Array<float>& arrayToScale, const float scaleInputToOutput, const float scaleNetToOutput, const cv::Size& producerSize) const
     {
         try
         {
@@ -22,7 +22,7 @@ namespace op
         }
     }
 
-    void KeyPointScaler::scale(std::vector<Array<float>>& arrayToScalesToScale, const double scaleInputToOutput, const double scaleNetToOutput, const cv::Size& producerSize) const
+    void KeyPointScaler::scale(std::vector<Array<float>>& arrayToScalesToScale, const float scaleInputToOutput, const float scaleNetToOutput, const cv::Size& producerSize) const
     {
         try
         {
@@ -31,27 +31,27 @@ namespace op
                 // InputResolution
                 if (mScaleMode == ScaleMode::InputResolution)
                     for (auto& arrayToScale : arrayToScalesToScale)
-                        scaleKeyPoints(arrayToScale, 1./scaleInputToOutput);
+                        scaleKeyPoints(arrayToScale, 1.f/scaleInputToOutput);
                 // NetOutputResolution
                 else if (mScaleMode == ScaleMode::NetOutputResolution)
                     for (auto& arrayToScale : arrayToScalesToScale)
-                        scaleKeyPoints(arrayToScale, 1./scaleNetToOutput);
+                        scaleKeyPoints(arrayToScale, 1.f/scaleNetToOutput);
                 // [0,1]
                 else if (mScaleMode == ScaleMode::ZeroToOne)
                 {
-                    const auto scale = 1./scaleInputToOutput;
-                    const auto scaleX = scale / ((double)producerSize.width - 1.);
-                    const auto scaleY = scale / ((double)producerSize.height - 1.);
+                    const auto scale = 1.f/scaleInputToOutput;
+                    const auto scaleX = scale / ((float)producerSize.width - 1.f);
+                    const auto scaleY = scale / ((float)producerSize.height - 1.f);
                     for (auto& arrayToScale : arrayToScalesToScale)
                         scaleKeyPoints(arrayToScale, scaleX, scaleY);
                 }
                 // [-1,1]
                 else if (mScaleMode == ScaleMode::PlusMinusOne)
                 {
-                    const auto scale = 2./scaleInputToOutput;
-                    const auto scaleX = (scale / ((double)producerSize.width - 1.));
-                    const auto scaleY = (scale / ((double)producerSize.height - 1.));
-                    const auto offset = -1.;
+                    const auto scale = 2.f/scaleInputToOutput;
+                    const auto scaleX = (scale / ((float)producerSize.width - 1.f));
+                    const auto scaleY = (scale / ((float)producerSize.height - 1.f));
+                    const auto offset = -1.f;
                     for (auto& arrayToScale : arrayToScalesToScale)
                         scaleKeyPoints(arrayToScale, scaleX, scaleY, offset, offset);
                 }
